fix beta issues
